+++ docs/changelog.rst
@@ -3,8 +3,9 @@ Change Log
 
 v0.4.1 - released: XXXXX
 ----------------------------
-* https://github.com/DigitalCampus/oppia-mobile-android/issues/283 - add optional job title and organisation to profile
+* Fix https://github.com/DigitalCampus/oppia-mobile-android/issues/283 - add optional job title and organisation to profile
 * Reset password facility https://github.com/DigitalCampus/oppia-mobile-android/issues/issue/292
+* Fix https://github.com/DigitalCampus/django-oppia/issues/173
 
 v0.4.0 - released: 1-Apr-2014
 ------------------------------